Renamed VectorFile to match the lack of adapter
